+++ R/param_surv_dist.R
@@ -1,10 +1,10 @@
 #' Parametric distributions for censored data
 #'
 #' @param values A character string of possible values. See `values_surv_dist`
-#'  in examples below.
+#'   in examples below.
 #'
 #' @details
-#' This parameter is used in `parsnip:::surv_reg()`.
+#' This parameter is used in `parsnip::survival_reg()`.
 #' @examples
 #' values_surv_dist
 #' surv_dist()
@@ -22,4 +22,3 @@ surv_dist <- function(values = values_surv_dist) {
 #' @export
 values_surv_dist <- c("weibull", "exponential", "gaussian",
                       "logistic", "lognormal", "loglogistic")
-