+++ tests/testthat/test_params.R
@@ -70,6 +70,7 @@ test_that('param ranges', {
   expect_equal(prior_terminal_node_coef(c(.1, .5))$range, list(lower = .1, upper = .5))
   expect_equal(prior_terminal_node_expo(c(1, 2))$range, list(lower = 1, upper = 2))
   expect_equal(prior_outcome_range(c(1, 5))$range, list(lower = 1, upper = 5))
+  expect_equal(num_knots(c(1, 5))$range, list(lower = 1, upper = 5))
 })
 
 
@@ -92,6 +93,6 @@ test_that('param values', {
   expect_equal(regularization_method()$values, dials:::values_regularization_method)
   expect_equal(diagonal_covariance(TRUE)$values, TRUE)
   expect_equal(summary_stat()$values, values_summary_stat)
-
+  expect_equal(survival_link()$values, values_survival_link)
 })